pyproject.toml
[project]
name = "rlmf-scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "polars>=1.30.0",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
    "streamlit-shortcuts>=0.2.1",
]
